refactor(racingcar): java컨벤션에 맞게 리팩토링

+++ src/main/java/racingcar/domain/Validator.java
@@ -6,8 +6,8 @@ import racingcar.domain.constant.ExceptionMessage;
 
 public class Validator {
 
-    private final String DELIMETER= ",";
-    private final int NAME_LENGTH_MAX=5;
+    private final String DELIMETER = ",";
+    private final int NAME_LENGTH_MAX = 5;
 
     public List<Car> validateCarName(String readCarName) {
         List<String> cars = List.of(readCarName.split(DELIMETER));
@@ -17,21 +17,22 @@ public class Validator {
     }
 
     private List<Car> convertStringToCar(List<String> cars) {
-        List<Car> convertCar = cars.stream().map(carName -> new Car(carName)).collect(Collectors.toList());
+        List<Car> convertCar = cars.stream().map(carName -> new Car(carName))
+            .collect(Collectors.toList());
         return convertCar;
     }
 
     private void validateNameLength(List<String> cars) {
-        if(cars.stream().filter(n -> n.length() > NAME_LENGTH_MAX).count() > 0){
+        if (cars.stream().filter(n -> n.length() > NAME_LENGTH_MAX).count() > 0) {
             throw new IllegalArgumentException(ExceptionMessage.ERROR_NAME_LENGTH.message());
         }
     }
 
-    public int validateInputTrial(String readTrial) {
-        int trial=0;
-        try{
+    public long validateInputTrial(String readTrial) {
+        int trial = 0;
+        try {
             trial = Integer.parseInt(readTrial);
-        }catch (Exception e){
+        } catch (Exception e) {
             throw new IllegalArgumentException(ExceptionMessage.ERROR_INPUT_TRIAL.message());
         }
         return trial;
